requirements.txt
pymongo
python-dotenv
azure-cosmos